[feat] 로직 구현

+++ app/src/main/java/org/sopt/certi/presentation/ui/resume/ResumeActivitiesScreen.kt
@@ -113,7 +113,7 @@ fun ResumeActivitiesRoute(
 fun ResumeActivitiesScreen(
     onNavigateToAdd: () -> Unit,
     resumeDataList: List<ActivityData>,
-    onDeleteClick: () -> Unit,
+    onDeleteClick: (Long) -> Unit,
     modifier: Modifier = Modifier
 ) {
     LazyColumn(
